pyproject.toml
[project]
name = "bank-product-monitoring-app"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "aiohttp>=3.13.2",
    "beautifulsoup4>=4.14.0",
    "httpx>=0.28.1",
    "lxml>=6.0.0",
    "matplotlib>=3.10.8",
    "numpy>=2.0.0",
    "pandas>=2.2.0",
    "pyqt6>=6.10.1",
    "qasync>=0.28.0",
    "playwright>=1.57.0",
    "playwright-stealth>=2.0.0",
    "fake-useragent>=1.5.1",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "bs4>=0.0.2",
    "sqlalchemy>=2.0.0",
    "psycopg2-binary>=2.9.0",
    "selenium>=4.15.0",
    "webdriver-manager>=4.0.0",
    "openpyxl>=3.1.0",
]
